third_party/blink/web_tests/external/wpt/web-nfc/NFCWriter_push.https-expected.txt

This is a testharness.js-based test.
PASS Test that promise is rejected with TypeError if NDEFMessageSource is invalid.
PASS Test that promise is rejected with SyntaxError if NDEFMessageSource contains invalid records.
PASS NFCWriter.push should fail if abort push request before push happends.
PASS NFCWriter.push should fail if signal's aborted flag is set.
PASS NFCWriter.push should fail if signal is not an AbortSignal.
PASS Synchronously signaled abort.
PASS NFCWriter.push should fail when NFC HW is disabled.
PASS NFCWriter.push should fail when NFC HW is not supported.
PASS NFCWriter.push should fail with TypeError when invalid timeout is provided.
PASS NFCWriter.push should fail with TypeError when invalid negative timeout value is provided.
PASS NFCWriter.push should fail with TimeoutError when timer expires.
PASS Reject promise with NotSupportedError if NFC message size exceeds 32KB.
PASS Reject promise with SyntaxError if WebNFC Id cannot be created from provided URL.
PASS Reject promise with exceptions thrown from serializing the 'json' record data.
PASS NFCWriter.push should fail with TypeError when invalid target value is provided.
PASS Test that WebNFC API is not accessible from iframe context.
PASS NFCWriter.push should succeed when NFC HW is enabled
PASS NFCWriter.push NDEFMessage containing text, json, opaque, url and external records with default NFCPushOptions.
PASS Test that NFCWriter.push succeeds when message is DOMString.
PASS Test that NFCWriter.push succeeds when message is ArrayBuffer.
PASS NFCWriter.push with 'empty' record should succeed.
PASS Check that default NFCPushOptions values are correctly set.
PASS Check that provided NFCPushOptions values are correctly converted.
PASS NFCWriter.push should read data when ignoreRead is false.
PASS NFCWriter.push should ignore reading data when ignoreRead is true.
PASS NFCWriter.push should replace all previously configured push operations.
PASS Test that recordType should be set to 'text' if NDEFRecordInit.record's recordType is undefined and NDEFRecordInit.record's data is DOMString.
PASS Test that recordType should be set to 'opaque' if NDEFRecordInit.record's recordType is undefined and NDEFRecordInit.record's data is ArrayBuffer.
PASS Test that recordType should be set to 'json' if NDEFRecordInit.record's recordType is undefined and NDEFRecordInit.record's data is not DOMString or ArrayBuffer.
PASS Test that mediaType should be set to 'text/plain' if NDEFRecordInit.record's recordType is 'text' and NDEFRecordInit.record's mediaType is undefined.
PASS Test that mediaType should be set to 'application/octet-stream' if NDEFRecordInit.record's recordType is 'opaque' and NDEFRecordInit.record's mediaType is undefined.
PASS Test that mediaType should be set to 'application/json' if NDEFRecordInit.record's recordType is 'json' and NDEFRecordInit.record's mediaType is undefined.
FAIL Test that mediaType should be set to 'text/plain' if NDEFRecordInit.record's recordType is 'url' and NDEFRecordInit.record's mediaType is undefined. promise_test: Unhandled rejection with value: object "TypeError: Cannot read property 'substring' of null"
PASS Test that mediaType should be set to 'application/octet-stream' if NDEFRecordInit.record's recordType is external type and NDEFRecordInit.record's mediaType is undefined.
Harness: the test ran to completion.
